Created User struct and the repository.go
The User struct is used to represent a User.

Added CreateUser function that create a user on the database

Added GetUserByUsername function that returns the pointer to a User that has the username

Added GetUserByID function that returns the pointer to a User that has the id.

Added ListUsers function that returns the list of users present in the db