Created Group struct and the repository.go
The Group struct is used to represent a Group.

Added CreateGroup function that puts a group in the database's collection

Added GetGroupByID function to get the group from the database using the id

Added ListGroups function to get all the groups in the database

Added AddMember function that adds a user to a group

Added RemoveMember function that removes a user from a group

Added RenameGroup function that removes a user from a group.